zerodha-dashboard-opus1/backend/requirements.txt
Flask==3.0.2
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
alembic==1.13.1
APScheduler==3.10.4
kiteconnect==5.0.1
cryptography==42.0.5
python-dotenv==1.0.1
pandas==2.2.1
numpy==1.26.4
requests==2.31.0

# Development
pytest==8.1.1
pytest-flask==1.3.0
black==24.2.0
flake8==7.0.0
